# Add precompilation for FourierTools