--- a/src/FourierTools.jl
+++ b/src/FourierTools.jl
@@ -8,6 +8,10 @@ using LinearAlgebra
 using IndexFunArrays
 using ChainRulesCore
 using NDTools
+
+using SnoopPrecompile
+
+
 @reexport using NFFT
 FFTW.set_num_threads(4)
 
@@ -30,5 +34,8 @@ include("correlations.jl")
 include("damping.jl")
 include("czt.jl")
 include("fractional_fourier_transform.jl")
+
+
+include("precompilation.jl")
 
 end # module