Some timings.

There is certainly improvement in Time-to-first-X (TTFX) but `using FourierTools` slowed down :/


```julia
# THIS PR

julia> @time @eval begin
        using FourierTools; resample(randn((3,3)), (4,4));
       end
  1.837442 seconds (4.37 M allocations: 330.662 MiB, 5.73% gc time, 37.63% compilation time: 35% of which was recompilation)
4×4 Matrix{Float64}:
  2.50649    0.177745  -1.26525     1.06349
  0.356397  -0.684275  -2.04213    -1.00146
 -1.33411   -0.982222  -0.792645   -1.14453
  0.815986  -0.120202  -0.0157615   0.920427



julia> @time @eval begin
        using FourierTools
       end
  1.233218 seconds (3.56 M allocations: 291.111 MiB, 7.55% gc time, 7.26% compilation time: 36% of which was recompilation)




# master branch

julia> @time begin
        using FourierTools
       end
  0.813568 seconds (2.46 M allocations: 178.026 MiB, 3.53% gc time, 9.22% compilation time: 25% of which was recompilation)

julia> @time @eval begin
        using FourierTools; resample(randn((3,3)), (4,4));
       end
  2.399415 seconds (8.97 M allocations: 544.981 MiB, 12.13% gc time, 69.04% compilation time: 1% of which was recompilation)
4×4 Matrix{Float64}:
 -0.206244   1.54404   2.20982   0.459536
  0.4502     1.25477   1.59361   0.789031
 -0.259531  -1.75128  -0.430379  1.06137
 -0.915975  -1.46202   0.185834  0.731878
```