@roflmaostc 2022-11-22
Add this copied message from slack for a moment:
```
Here's what I would do:
First, make a "representative workload" for your package. Something that uses the most common latency-sensitive functions. This is probably what you have inside your SnoopPrecompile block?
Put this workload inside a function, then check it's type stable with JET's report_opt . Make type stable as necessary
Check for type piracy. This can be automated using this script here: https://discourse.julialang.org/t/piracy-hunting/90119. Remove all piracy you find
```