@roflmaostc 2023-03-21
Julia 1.9:

```
# without precomp
julia> @time @eval begin
               using FourierTools; resample(randn((3,3)), (4,4));
              end
  2.347064 seconds (7.83 M allocations: 521.891 MiB, 7.21% gc time, 71.20% compilation time)


julia> @time @eval begin
               using FourierTools
              end
  0.667420 seconds (1.51 M allocations: 101.459 MiB, 6.05% gc time, 0.93% compilation time)




# with precompile

julia> @time @eval begin
               using FourierTools
              end

  0.896489 seconds (2.05 M allocations: 141.270 MiB, 6.69% gc time, 0.69% compilation time)


julia> @time @eval begin
               using FourierTools; resample(randn((3,3)), (4,4));
              end
  1.032369 seconds (2.35 M allocations: 160.550 MiB, 6.08% gc time, 13.95% compilation time)
4×4 Matrix{Float64}:
 1.40546     -1.01361   -0.690086  1.72899
 0.0448578   -1.07521   -0.575482  0.544589
 0.00392657  -0.224145   1.05394   1.28201
 1.36453     -0.162546   0.939333  2.46641

``